HTML Alert Message Box
Alert message box using HTML, CSS and JavaScript.

We use alert messages to inform users about something. In this example, we have Info, Successful, Warning and Error messages. It is easy to add more.

Notice in the HTML, on line 72, the class is 'g2s-alert-i, this is the informational box. For the warning box, change it to 'g2s-alert-w'.

g2s-alert-i = Informational
g2s-alert-s = Successful
g2s-alert-w = Warning
g2s-alert-e = Error